development version corrected
I reverted the merge pull request because I wanted to do a rebase rather than a merge commit... but it didn't work: it reverted also the development version to the master state. Luckily I had made a local save. So now I commit eveything back to the master branch. Next time I'll try to select the good type of commit