Started debugging rSoilNitrogen, updated allvariables.xlsx
By executing an example run, seeing where it fails, where there are missing or unexpected values, etc...
As a result: fixed fComputeRNforMineralization and fComputeNMineralization to be compatible with vector/matrix/data.frame inputs.

NOTE: I made it way more it complicated than it should be, I will change it for the next commit.

TO-DO: Change with ifelse and mat[condition] <- value. Also do this for every function of this type, for instance, next is "fComputeNDrainage"