Soil nitrogen: 1st working version
Working as in : simulation can run and not break for supposedly ever (700+ days at least) and few unexpected values.
Only NAVL (and thus SNAVL) is known to take negative values somehow. It's very likely a problem with the timestep (value to WHAT point of the timestep should we consider?) ... I will investigate.

Updated changelog, although I haven't been so rigorous on keeping notes for the past few days (a lot of small tweaks).